Add /api/about endpoint and supporting register-descriptor schema
Schema changes laying the groundwork for a complete UPS register dump:

- New ups_reg_category_t (CONFIG / MEASUREMENT / IDENTITY / DIAGNOSTIC) on
  every config descriptor; defaults to CONFIG so existing entries are
  unchanged
- New UPS_CFG_FLAGS type for multi-bit registers (status bitfields, error
  registers) where many bits can be set simultaneously
- has_sentinel / sentinel_value on each descriptor so drivers can tag
  "not applicable on this firmware/SKU" reads (e.g. SMT bypass voltage
  0xFFFF, load-shed config 0x0000) for "N/A" rendering
- raw_value widened uint16 -> uint32 throughout (including
  ups_bitfield_opt_t.value) so 32-bit registers like UPSStatus and
  PowerSystemError surface their high bits instead of silently
  truncating
- sku[33] field on ups_inventory_t for SKU/part number

API:
- /api/about returns the union of inventory and every config descriptor
  the driver exposes, regardless of category. Replaces the implicit
  "everything is a config setting" assumption with explicit categorisation.
- /api/config/ups now filters server-side to category=CONFIG so the
  operator-facing settings page doesn't get flooded once the descriptor
  table balloons.
- /api/status emits inventory.sku alongside the existing fields.

Plumbing:
- monitor's config snapshot loop skips non-CONFIG descriptors so the
  ups_config audit table doesn't get flooded with telemetry-rate
  measurements.
- driver-api.md: drop the stale "SMT is a stub" rough-edge note, refresh
  the resolve_config_regs paragraph (SMT now uses it), and update the
  new-driver checklist example away from SMT.

+++ tests/test_config_validation.c
@@ -31,7 +31,7 @@ static int fake_config_write(void *transport, const ups_config_reg_t *reg, uint1
 }
 
 static int fake_config_read(void *transport, const ups_config_reg_t *reg,
-                            uint16_t *raw, char *str, size_t str_sz)
+                            uint32_t *raw, char *str, size_t str_sz)
 {
     (void)transport; (void)reg; (void)str; (void)str_sz;
     if (raw) *raw = 0;
